+++ extensions/contract-termination/src/main/java/de/sovity/edc/extension/contacttermination/ContractAgreementTerminationService.java
@@ -16,15 +16,20 @@ package de.sovity.edc.extension.contacttermination;
 
 import de.sovity.edc.extension.contacttermination.query.ContractAgreementTerminationDetailsQuery;
 import de.sovity.edc.extension.contacttermination.query.TerminateContractQuery;
+import de.sovity.edc.extension.messenger.SovityMessage;
 import de.sovity.edc.extension.messenger.SovityMessenger;
+import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import lombok.val;
 import org.eclipse.edc.spi.EdcException;
 import org.eclipse.edc.spi.monitor.Monitor;
+import org.eclipse.edc.spi.observe.Observable;
+import org.eclipse.edc.spi.observe.ObservableImpl;
 import org.jetbrains.annotations.Nullable;
 import org.jooq.DSLContext;
 
 import java.time.OffsetDateTime;
+import java.util.function.Consumer;
 
 import static de.sovity.edc.ext.db.jooq.enums.ContractTerminatedBy.COUNTERPARTY;
 import static de.sovity.edc.ext.db.jooq.enums.ContractTerminatedBy.SELF;
@@ -37,14 +42,19 @@ public class ContractAgreementTerminationService {
     private final TerminateContractQuery terminateContractQuery;
     private final Monitor monitor;
     private final String thisParticipantId;
+    @Getter
+    private final Observable<ContractTerminationObserver> contractTerminationObservable = new ObservableImpl<>();
 
     /**
      * This is to terminate an EDC's own contract.
      * If the termination comes from an external system, use
-     * {@link #terminateCounterpartyAgreement(DSLContext, String, ContractTerminationParam)}
+     * {@link #terminateAgreementAsCounterparty(DSLContext, String, ContractTerminationParam)}
      * to validate the counter-party's identity.
      */
     public OffsetDateTime terminateAgreementOrThrow(DSLContext dsl, ContractTerminationParam termination) {
+
+        val starterEvent = ContractTerminationEvent.from(termination, OffsetDateTime.now(), thisParticipantId);
+        notifyObservers(it -> it.contractTerminationStartedFromThisInstance(starterEvent));
 
         val details = contractAgreementTerminationDetailsQuery.fetchAgreementDetailsOrThrow(dsl, termination.contractAgreementId());
 
@@ -58,16 +68,22 @@ public class ContractAgreementTerminationService {
 
         val terminatedAt = terminateContractQuery.terminateConsumerAgreementOrThrow(dsl, termination, SELF);
 
+        val endEvent = ContractTerminationEvent.from(termination, terminatedAt, thisParticipantId);
+        notifyObservers(it -> it.contractTerminationCompletedOnThisInstance(endEvent));
+
         notifyTerminationToProvider(details.counterpartyAddress(), termination);
 
         return terminatedAt;
     }
 
-    public OffsetDateTime terminateCounterpartyAgreement(
+    public OffsetDateTime terminateAgreementAsCounterparty(
         DSLContext dsl,
         @Nullable String identity,
         ContractTerminationParam termination
     ) {
+        val starterEvent = ContractTerminationEvent.from(termination, OffsetDateTime.now(), null);
+        notifyObservers(it -> it.contractTerminatedByCounterpartyStarted(starterEvent));
+
         val details = contractAgreementTerminationDetailsQuery.fetchAgreementDetailsOrThrow(dsl, termination.contractAgreementId());
 
         if (details == null) {
@@ -90,15 +106,35 @@ public class ContractAgreementTerminationService {
 
         val agent = thisParticipantId.equals(details.counterpartyId()) ? SELF : COUNTERPARTY;
 
-        return terminateContractQuery.terminateConsumerAgreementOrThrow(dsl, termination, agent);
+        val result = terminateContractQuery.terminateConsumerAgreementOrThrow(dsl, termination, agent);
+
+        val endEvent = ContractTerminationEvent.from(termination, OffsetDateTime.now(), details.counterpartyId());
+        notifyObservers(it -> it.contractTerminatedByCounterparty(endEvent));
+
+        return result;
     }
 
     public void notifyTerminationToProvider(String counterPartyAddress, ContractTerminationParam termination) {
+
+        val notificationEvent = ContractTerminationEvent.from(termination, OffsetDateTime.now(), null);
+        notifyObservers(it -> it.contractTerminationOnCounterpartyStarted(notificationEvent));
+
         sovityMessenger.send(
+            SovityMessage.class,
             counterPartyAddress,
             new ContractTerminationMessage(
                 termination.contractAgreementId(),
                 termination.detail(),
                 termination.reason()));
+    }
+
+    private void notifyObservers(Consumer<ContractTerminationObserver> call) {
+        for (val listener : contractTerminationObservable.getListeners()) {
+            try {
+                call.accept(listener);
+            } catch (Exception e) {
+                monitor.warning("Failure when notifying the contract termination listener.", e);
+            }
+        }
     }
 }